requirements.txt
appdirs==1.4.4
Babel==2.9.1
boolean.py==4.0
Brlapi==0.8.2
cffi==1.14.5
chardet==4.0.0
click==8.0.3
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.18
distlib==0.3.2
distro==1.5.0
docutils==0.16
file-magic==0.4.0
filelock==3.7.1
flameprof==0.4
gpg==1.15.1
idna==2.10
Jinja2==2.11.3
libcomps==0.1.18
lxml==4.6.5
MarkupSafe==1.1.1
mercurial==6.1.4
nftables==0.1
numpy==1.20.1
packaging==20.9
perf==0.1
pexpect==4.8.0
platformdirs==2.5.4
ply==3.11
podman-compose==1.0.6
ptyprocess==0.6.0
pycairo==1.20.1
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pyelftools==0.30
pyenchant==3.2.0
PyGObject==3.40.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python-linux-procfs==0.7.0
pytz==2021.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
rpm==4.16.1.3
scour==0.38.1
selinux==3.5
sepolicy==3.5
setools==4.4.3
six==1.15.0
sos==4.6.1
SSSDConfig==2.9.1
systemd-python==234
urllib3==1.26.5
virtualenv==20.21.1